Console and file extensions
