--- src/LobAccelerator.Client/Extensions/FileFormatExtensions.cs
@@ -1,10 +1,0 @@
-﻿namespace LobAccelerator.Client.Extensions
-{
-    public static class FileFormatExtensions
-    {
-        public static bool IsYaml(this string fileExtension)
-        {
-            return fileExtension == ".yaml" || fileExtension == ".yml";
-        }
-    }
-}
