+++ src/LobAccelerator.Client/Program.cs
@@ -1,5 +1,6 @@
 ﻿using CommandLine;
 using LobAccelerator.Client.Extensions;
+using LobAccelerator.Client.Models;
 using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Newtonsoft.Json;
 using System;
@@ -24,33 +25,14 @@ namespace LobAccelerator.Client
             }
             catch (Exception ex)
             {
-                DisplayErrorOnConsole(ex);
+                ConsoleExtensions.DisplayError(ex);
             }
 
 #if DEBUG
             Console.ReadLine();
 #endif
         }
-
-        private static void DisplayErrorOnConsole(Exception ex)
-        {
-            Console.ForegroundColor = ConsoleColor.Red;
-            Console.WriteLine("Something went wrong!");
-            Console.WriteLine($"Message: {ex.Message}");
-        }
-
-        private static void DisplayInfoMessage(string message)
-        {
-            Console.ForegroundColor = ConsoleColor.Blue;
-            Console.WriteLine(message);
-        }
-
-        private static void DisplaySuccessMessage(string message)
-        {
-            Console.ForegroundColor = ConsoleColor.Green;
-            Console.WriteLine(message);
-        }
-
+        
         private static void RunOptions(Options options)
         {
             RunOptionAsync(options).GetAwaiter().GetResult();
@@ -67,21 +49,19 @@ namespace LobAccelerator.Client
             string clientId = configuration["AzureAd:ClientId"];
             string url = configuration["LobEngine:Endpoint"];
 
-            DisplayInfoMessage("Input validated...");
+            ConsoleExtensions.DisplayInfoMessage("Input validated...");
 
-            var token = await GetTokenViaCode(resource, clientId);
-            string accessToken = token.AccessToken;
+            var accessToken = await ConsoleExtensions.GetTokenByCode(resource, clientId);
 
-            DisplayInfoMessage("Token acquired...");
+            ConsoleExtensions.DisplayInfoMessage("Token acquired...");
+            ConsoleExtensions.DisplayInfoMessage("Sending request...");
 
-            DisplayInfoMessage("Sending request...");
+            var manager = new LobManager();
+            await manager.ProvisionResourcesAsync(url, accessToken, files);
 
-            await SendFilesToEngine(url, accessToken, files);
-
-            DisplaySuccessMessage("Finished!");
+            ConsoleExtensions.DisplaySuccessMessage("Finished!");
         }
-
-
+        
         private static void ValidateInput(Options options)
         {
             if (!options.DefinitionsFiles.All(f => File.Exists(f)))
@@ -90,60 +70,7 @@ namespace LobAccelerator.Client
             if (!File.Exists(options.ConfigurationFile))
                 throw new FileNotFoundException("The configuration file was not found.");
         }
-
-        static async Task<AuthenticationResult> GetTokenViaCode(string resource, string clientId)
-        {
-            var ctx = new AuthenticationContext("https://login.microsoftonline.com/common");
-
-            DeviceCodeResult codeResult = await ctx.AcquireDeviceCodeAsync(resource, clientId);
-            Console.ResetColor();
-            Console.WriteLine("You need to sign in.");
-            Console.WriteLine("Message: " + codeResult.Message);
-
-            return await ctx.AcquireTokenByDeviceCodeAsync(codeResult);
-        }
-
-        static async Task SendFilesToEngine(string url, string accessToken, IEnumerable<string> files)
-        {
-            using (var httpClient = new HttpClient())
-            {
-                foreach (var file in files)
-                {
-                    var content = await GetFileContentAsync(file);
-                    var body = new StringContent(content, Encoding.UTF8, "application/json");
-
-                    httpClient.DefaultRequestHeaders.Add("X-Authorization", $"bearer {accessToken}");
-
-                    var response = await httpClient.PostAsync(url, body);
-                    response.EnsureSuccessStatusCode();
-                }
-            }
-
-        }
-
-        private static async Task<string> GetFileContentAsync(string file)
-        {
-            var fileExtension = Path.GetExtension(file);
-            var fileContent = await File.ReadAllTextAsync(file);
-
-            return fileExtension.IsYaml()
-                ? ConvertYamlToJson(fileContent)
-                : fileContent;
-        }
-
-        private static string ConvertYamlToJson(string fileContent)
-        {
-            var reader = new StringReader(fileContent);
-
-            var deserializer = new Deserializer();
-            var yamlObject = deserializer.Deserialize(reader);
-
-            var serializer = new JsonSerializer();
-            var stringWriter = new StringWriter();
-
-            serializer.Serialize(stringWriter, yamlObject);
-
-            return stringWriter.ToString();
-        }
+        
+        
     }
 }
